CU-86a69430z-NEON3 - Enable use of right-click to paste addresses

+++ tests/e2e/index.ts
@@ -15,9 +15,7 @@ export const launch = async (shouldResetStorage = true) => {
 
   const window = await electronApp.firstWindow()
 
-  if (shouldResetStorage) {
-    await window.evaluate('window.localStorage.clear()')
-  }
+  if (shouldResetStorage) await window.evaluate('window.localStorage.clear()')
 
   await window.reload({ waitUntil: 'load' })
